fix a C4244 x64 warning (which was repeated a lot of times)
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4666 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/DSUtil/text.h
@@ -11,7 +11,7 @@ T Explode(T str, CAtlList<T>& sl, SEP sep, size_t limit = 0)
 {
 	sl.RemoveAll();
 
-	for (ptrdiff_t i = 0, j = 0; ; i = j+1) {
+	for (int i = 0, j = 0; ; i = j+1) {
 		j = str.Find(sep, i);
 
 		if (j < 0 || sl.GetCount() == limit-1) {
